% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{download_gwas_files}
\alias{download_gwas_files}
\title{Download GWAS UK Biobank GWAS Results}
\usage{
download_gwas_files(dest = getwd(), ...)
}
\arguments{
\item{dest}{Local directory where the files will be downloaded to.}

\item{...}{Additional arguments passed to `base::download.file()`.}
}
\description{
Download GWAS UK Biobank GWAS Results
}
