% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides.R
\name{slides}
\alias{slides}
\title{Display the tutorial slides}
\usage{
slides()
}
\value{
The return code from calling the viewer function
}
\description{
Display the included pdf slides
}
\details{
The function relies on the internal function \code{pdf_viewer}
which we are borrowing with grateful acknowledgement from R Core
and the \code{RShowDoc} function.
}
\author{
Dirk Eddelbuettel
}
